# spec/swagger.yaml
definitions:
...
  channel.ChannelCapabilities:
    properties:
      attachments:
        type: boolean
      block_streaming:
        type: boolean
      buttons:
        type: boolean
      chat_types:
        items:
          type: string
        type: array
      edit:
        type: boolean
      markdown:
        type: boolean
      media:
        type: boolean
      native_commands:
        type: boolean
      polls:
        type: boolean
      reactions:
        type: boolean
      reply:
        type: boolean
      rich_text:
        type: boolean
      streaming:
        type: boolean
      text:
        type: boolean
      threads:
        type: boolean
      unsend:
        type: boolean
    type: object
  channel.ChannelConfig:
    properties:
      bot_id:
        type: string
      channel_type:
        type: string
      created_at:
        type: string
      credentials:
        additionalProperties: {}
        type: object
      external_identity:
        type: string
      id:
        type: string
      routing:
        additionalProperties: {}
        type: object
      self_identity:
        additionalProperties: {}
        type: object
      status:
        type: string
      updated_at:
        type: string
      verified_at:
        type: string
    type: object
  channel.ChannelIdentityBinding:
    properties:
      channel_identity_id:
        type: string
      channel_type:
        type: string
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  channel.ConfigSchema:
    properties:
      fields:
        additionalProperties:
          $ref: '#/definitions/channel.FieldSchema'
        type: object
      version:
        type: integer
    type: object
  channel.FieldSchema:
    properties:
      description:
        type: string
      enum:
        items:
          type: string
        type: array
      example: {}
      required:
        type: boolean
      title:
        type: string
      type:
        $ref: '#/definitions/channel.FieldType'
    type: object
  channel.FieldType:
    enum:
    - string
    - secret
    - bool
    - number
    - enum
    type: string
    x-enum-varnames:
    - FieldString
    - FieldSecret
    - FieldBool
    - FieldNumber
    - FieldEnum
...
  channel.TargetHint:
    properties:
      example:
        type: string
      label:
        type: string
    type: object
  channel.TargetSpec:
    properties:
      format:
        type: string
      hints:
        items:
          $ref: '#/definitions/channel.TargetHint'
        type: array
    type: object
...
  channel.UpsertChannelIdentityConfigRequest:
    properties:
      config:
        additionalProperties: {}
        type: object
    type: object
  channel.UpsertConfigRequest:
    properties:
      credentials:
        additionalProperties: {}
        type: object
      external_identity:
        type: string
      routing:
        additionalProperties: {}
        type: object
      self_identity:
        additionalProperties: {}
        type: object
      status:
        type: string
      verified_at:
        type: string
    type: object
...
  handlers.ChannelMeta:
    properties:
      capabilities:
        $ref: '#/definitions/channel.ChannelCapabilities'
      config_schema:
        $ref: '#/definitions/channel.ConfigSchema'
      configless:
        type: boolean
      display_name:
        type: string
      target_spec:
        $ref: '#/definitions/channel.TargetSpec'
      type:
        type: string
      user_config_schema:
        $ref: '#/definitions/channel.ConfigSchema'
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      access_token:
        type: string
      display_name:
        type: string
      expires_at:
        type: string
      role:
        type: string
      token_type:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
  /auth/login:
    post:
      description: Validate user credentials and issue a JWT
      parameters:
      - description: Login request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /bots/{id}/channel/{platform}:
    get:
      description: Get bot channel configuration
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel platform
        in: path
        name: platform
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/channel.ChannelConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get bot channel config
      tags:
      - bots
    put:
      description: Update bot channel configuration
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel platform
        in: path
        name: platform
        required: true
        type: string
      - description: Channel config payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/channel.UpsertConfigRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/channel.ChannelConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update bot channel config
      tags:
      - bots
...
  /channels:
    get:
      description: List channel meta information including capabilities and schemas
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ChannelMeta'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List channel capabilities and schemas
      tags:
      - channel
  /channels/{platform}:
    get:
      description: Get channel meta information including capabilities and schemas
      parameters:
      - description: Channel platform
        in: path
        name: platform
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ChannelMeta'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get channel capabilities and schemas
      tags:
      - channel
...
  /users/me/channels/{platform}:
    get:
      description: Get channel binding configuration for current user
      parameters:
      - description: Channel platform
        in: path
        name: platform
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/channel.ChannelIdentityBinding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get channel user config
      tags:
      - channel
    put:
      description: Update channel binding configuration for current user
      parameters:
      - description: Channel platform
        in: path
        name: platform
        required: true
        type: string
      - description: Channel user config payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/channel.UpsertChannelIdentityConfigRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/channel.ChannelIdentityBinding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update channel user config
      tags:
      - channel
...
swagger: "2.0"